requirements.txt
pytest==2.7.2
pytest-describe==0.10.3
pyhamcrest==1.9.0
coverage==4.0.3
